Refactor and update copy on CriticalCssStatus
